requirements.txt
flask
pillow
opencv-python
numpy
keras
tensorflow
sklearn
requests